services/app/requirements.txt
fastapi==0.115.0
pydantic==2.9.2
uvicorn[standard]==0.30.6
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.12
langchain-openai==0.1.14
chromadb==1.2.0
python-dotenv==1.0.0
requests==2.31.0
pymongo==4.6.1
motor==3.3.2
